requirements.txt
accelerate==0.21.0
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
diffusers==0.18.2
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
pyre-extensions==0.0.29
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
replicate==0.9.0
requests==2.31.0
six==1.16.0
sympy==1.12
torch=2.0.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xformers==0.0.20
zipp==3.16.2
